requirements.txt
telebot
librosa
soundfile
pillow
roman